Fixed ignored hostgroup when colsep is specified.
Added an optional $parent argument to the Arg constructor; the
hostgroups of the parent object are preserved.

Added a hostgroups accessor to Arg package.

The host group name cannot contain '@' characters, and must
appear once per line, in the end of it.

The splitting of the arguments during parsing preserves the
hostgroup by inheriting this from the current "parent" object.